Produce useful output
Each specified transaction type results in a TSV file summarizing
transactions of that type. Some minor improvements to log messages.

In aggregate_transactions_by_type, call set_index() on the initial
DataFrame with the name of the field that will later be used to join
each aggregate DataFrame. Call to_frame() on Series objects that are
returned by the aggregation functions, which allows non-index columns to
be renamed (necessary to avoid collisions). Add the index field to the
'retval' DataFrame on creation, otherwise an error will occur because
'retval' lacks the index and can't be joined to the DataFrames in
'result'. Use an outer join because we should assume we might need to
deal with more or less sparse data.